requirements.txt
flask==3.1.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.1